+++ src/history/createHashHistory.js
@@ -1,16 +1,13 @@
-import { Observable } from 'rxjs/Observable';
+import { fromEvent } from 'rxjs/observable/fromEvent';
 
-import 'rxjs/add/observable/fromEvent';
-import 'rxjs/add/observable/of';
-import 'rxjs/add/operator/first';
-import 'rxjs/add/operator/take';
-import 'rxjs/add/operator/startWith';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/publishReplay';
+import { startWith } from 'rxjs/operators/startWith';
+import { map } from 'rxjs/operators/map';
+import { publishReplay } from 'rxjs/operators/publishReplay';
+import { refCount } from 'rxjs/operators/refCount';
 
-import { locationFromUrl } from './utils/locationFromUrl';
+import { locationFromUrl } from '../utils/locationFromUrl';
 
-import { splitUrl } from './utils/splitUrl';
+import { splitUrl } from '../utils/splitUrl';
 
 export function createHashHistory() {
   function currentLocation(source) {
@@ -31,11 +28,12 @@ export function createHashHistory() {
   let replace;
 
   if ('onpopstate' in window) {
-    location = Observable.fromEvent(window, 'popstate')
-      .map(() => currentLocation('pop'))
-      .startWith(currentLocation('init'))
-      .publishReplay(1)
-      .refCount();
+    location = fromEvent(window, 'popstate').pipe(
+      map(() => currentLocation('pop')),
+      startWith(currentLocation('init')),
+      publishReplay(1),
+      refCount()
+    );
 
     push = (url, state = null, title = null) => {
       window.history.pushState(state, title, url);
@@ -51,11 +49,12 @@ export function createHashHistory() {
       window.location.assign(url);
     };
 
-    location = Observable.fromEvent(window, 'hashchange')
-      .map(() => currentLocation('pop'))
-      .startWith(currentLocation('init'))
-      .publishReplay(1)
-      .refCount();
+    location = fromEvent(window, 'hashchange').pipe(
+      map(() => currentLocation('pop')),
+      startWith(currentLocation('init')),
+      publishReplay(1),
+      refCount()
+    );
   }
 
   return {