+++ src/index.js
@@ -1,6 +1,6 @@
 export { Router } from './Router';
-export { createServerHistory } from './createServerHistory';
-export { createBrowserHistory } from './createBrowserHistory';
-export { createHashHistory } from './createHashHistory';
-export { createTestHistory } from './createTestHistory';
+export { createServerHistory } from './history/createServerHistory';
+export { createBrowserHistory } from './history/createBrowserHistory';
+export { createHashHistory } from './history/createHashHistory';
+export { createTestHistory } from './history/createTestHistory';
 export { RouteCollection } from './RouteCollection';